+++ backend/.env.example
@@ -1,3 +1,4 @@
 GITHUB_TOKEN=
 OPENAI_API_KEY=
 GITHUB_API_URL=https://api.github.com/graphql
+REPOS_PATH=./repos_cache
